Fixed broken layout on Nexus 6P.
Layout was fine in Chrome's device mode (the mobile emulator) but on the actual Nexus 6P the 'Medium' and 'Twitter' links had disappeared off the edge of the screen. Decided to use floats instead of inline blocks so that I could make use of a technique that uses auto width to take up the rest of the available space.

http://stackoverflow.com/questions/9635347/fixing-sub-pixel-rounding-issue-in-a-css-fluid-grid