Removed 100% height of sections.
The difficulty with 100% section height is ensuring that all content fits within that height on all screen sizes. The reason for this choice originally was because I thought it made sense in combination with scroll snapping that I intend to implement later. However I do not think this potential slight improvement to the UX is worth the extra effort of ensuring content is displayed correctly and risk of damaging UX if it is shown incorrectly for some users.